fix(ui): split About link constants and harden AboutSection test selectors
