+++ frontend/src/components/settings/aboutLinks.ts
@@ -1,0 +1,11 @@
+export const SOURCE_URL = 'https://github.com/studio-saelix/sencho';
+export const LICENSE_URL = 'https://github.com/studio-saelix/sencho/blob/main/LICENSE';
+export const LICENSING_DOCS_URL = 'https://docs.sencho.io/features/licensing';
+export const CHANGELOG_URL = 'https://github.com/studio-saelix/sencho/blob/main/CHANGELOG.md';
+
+export const ABOUT_LINK_URLS = {
+    source: SOURCE_URL,
+    license: LICENSE_URL,
+    licensingDocs: LICENSING_DOCS_URL,
+    changelog: CHANGELOG_URL,
+} as const;
